Bootstrap-estimate of Flux-error

Bootstrap-estimate of Flux-error

Usage

bootstrap_error(
  x,
  n_samples = 50,
  sd_x_ppm = NULL,
  n_replicates = NULL,
  sample_from = "gasdata",
  rep_cols = NULL
)

# S3 method for class 'cfp_pfres'
bootstrap_error(
  x,
  n_samples = 50,
  sd_x_ppm = NULL,
  n_replicates = NULL,
  sample_from = "gasdata",
  rep_cols = NULL
)

make_bootstrap_model(
  x,
  n_samples = 50,
  sd_x_ppm = NULL,
  n_replicates = NULL,
  sample_from = "gasdata",
  rep_cols = NULL
)

# S3 method for class 'cfp_pfres'
make_bootstrap_model(
  x,
  n_samples = 50,
  sd_x_ppm = NULL,
  n_replicates = NULL,
  sample_from = "gasdata",
  rep_cols = NULL
)

calculate_bootstrap_error(x, y)

# S3 method for class 'cfp_pfres'
calculate_bootstrap_error(x, y)

Arguments

x

A cfp_pfres model result from a call to pro_flux().

n_samples

The number of samples to take in the bootstrapping.

sd_x_ppm

An optional estimate of the standard deviation of x_ppm. Can be either

a single value applied equally to all

a data.frame with a column of the same name that maps a value to every observation depth. See depth_structure() for an easy way to create it.

be provided as its own column already present in x$gasdata.

n_replicates

The number of replicates to be generated if sd_x_ppm is set.

sample_from

From which dataset to sample the bootstrapping dataset. Can either be 'gasdata' or 'soilphys' or 'both'.

rep_cols

The id_cols that represent repetitions. If removed, the repetitions in soilphys of each profile must match in their structure exactly.

y

The result of the bootstrap model.

Value

x with added columns DELTA_flux and DELTA_prod as an estimate of the error of of the corresponding columns in the same units.

General procedure

bootstrap_error() is mostly a wrapper around two functions that can also be run separately.

In make_bootstrap_model(), for sample_from = "gasdata" the gasdata concentration data is resampled for every depth and profile a total number of n_samples. This is done by randomly sampling the observations at each depth without changing the number of observations but while allowing replacing. If rep_cols are given, these columns are removed from the id_cols and the resulting profiles combined as one.

For sample_from = "soilphys", the soilphys data is combined using the rep_cols as repetitions. Among every remaining profile and depth, one observation across all repetitions is chosen for each of n_samples. sample_from = "both" applies both methods above. Each newly sampled profile is identifiable by the added bootstrap_id column which is also added to id_cols.

After this new model is run again, the bootstap error is caculated in calculate_bootstrap_error(). This is the standard deviation of the production and flux parameters across all bootstrapped model runs and is calculated for each profile and layer of the original model, or for each destinct profile in the new model without rep_cols. These are returned together with the mean values of prod, flux and F0 across all runs in the PROFLUX data.frame and can thereby be extracted by efflux() and production().

Artificial observations in gasdata

If there are not enough observations per depth (e.g.) because there is only one measurement per depth, it is possible to create artificial observations by providing n_replicates and sd_x_ppm. Here, every depth of every profile is first averaged to its mean (redundant if there is only one observation). Then, a random dataset of n_replicates observations is generated that is normally distributed around the mean with a standard deviation (in ppm) of sd_x_ppm. These observations are then resampled as described above. Note that this error should be representative of the sampling error in the field and not the measurement error of the measurement device, which is much lower.

Examples

PROFLUX <- pro_flux(ConFluxPro::base_dat)
PROFLUX_BSE <- bootstrap_error(PROFLUX)
#> 
#> validating datasets
#> id_cols: site, Date, gas, bootstrap_id
#> 1200 unique profiles
efflux(PROFLUX_BSE)
#> 
#> A cfp_profile object 
#> id_cols: site Date gas 
#> 24  unique profiles 
#> 
#>      site       Date gas     efflux prof_id DELTA_efflux
#> 1  site_a 2021-01-01 CO2  0.8429610       1   0.02633798
#> 2  site_b 2021-01-01 CO2  0.6763911       2   0.01389758
#> 3  site_a 2021-02-01 CO2  1.3784963       3   0.02182533
#> 4  site_b 2021-02-01 CO2  1.0666943       4   0.01496520
#> 5  site_a 2021-03-01 CO2  1.7424766       5   0.03833253
#> 6  site_b 2021-03-01 CO2  1.3910526       6   0.01160696
#> 7  site_a 2021-04-01 CO2  4.1396328       7   0.04253622
#> 8  site_b 2021-04-01 CO2  3.3456025       8   0.05015449
#> 9  site_a 2021-05-01 CO2  5.9676759       9   0.08219772
#> 10 site_b 2021-05-01 CO2  4.4710172      10   0.05518122
#> 11 site_a 2021-06-01 CO2  7.2678452      11   0.09628114
#> 12 site_b 2021-06-01 CO2  5.6783657      12   0.07184852
#> 13 site_a 2021-07-01 CO2 10.1469971      13   0.14381006
#> 14 site_b 2021-07-01 CO2  8.5688079      14   0.10590895
#> 15 site_a 2021-08-01 CO2  8.0515776      15   0.17054155
#> 16 site_b 2021-08-01 CO2  6.9149107      16   0.10510296
#> 17 site_a 2021-09-01 CO2  6.2443430      17   0.12090202
#> 18 site_b 2021-09-01 CO2  5.2440922      18   0.07166082
#> 19 site_a 2021-10-01 CO2  3.9959103      19   0.12215896
#> 20 site_b 2021-10-01 CO2  3.2447621      20   0.04116001
#> 21 site_a 2021-11-01 CO2  1.7529403      21   0.03401804
#> 22 site_b 2021-11-01 CO2  1.5550672      22   0.02664799
#> 23 site_a 2021-12-01 CO2  1.2077043      23   0.02362541
#> 24 site_b 2021-12-01 CO2  1.0862780      24   0.01307257

PROFLUX_BSE <- bootstrap_error(PROFLUX, n_replicates = 5, sd_x_ppm = 25)
#> Adding gasdata replicates as normal distribution of 5 samples around the mean.
#> 
#> validating datasets
#> id_cols: site, Date, gas, bootstrap_id
#> 1200 unique profiles
efflux(PROFLUX_BSE)
#> 
#> A cfp_profile object 
#> id_cols: site Date gas 
#> 24  unique profiles 
#> 
#>      site       Date gas     efflux prof_id DELTA_efflux
#> 1  site_a 2021-01-01 CO2  0.8305233       1   0.06948057
#> 2  site_b 2021-01-01 CO2  0.6892845       2   0.02745311
#> 3  site_a 2021-02-01 CO2  1.3804727       3   0.08592559
#> 4  site_b 2021-02-01 CO2  1.0707094       4   0.03843973
#> 5  site_a 2021-03-01 CO2  1.7922332       5   0.05939070
#> 6  site_b 2021-03-01 CO2  1.4254358       6   0.05062181
#> 7  site_a 2021-04-01 CO2  3.9879387       7   0.07492112
#> 8  site_b 2021-04-01 CO2  3.3984600       8   0.05472674
#> 9  site_a 2021-05-01 CO2  5.9539752       9   0.11643840
#> 10 site_b 2021-05-01 CO2  4.4455024      10   0.05375143
#> 11 site_a 2021-06-01 CO2  7.2869279      11   0.06354410
#> 12 site_b 2021-06-01 CO2  5.7311535      12   0.12859636
#> 13 site_a 2021-07-01 CO2 10.2930064      13   0.09670892
#> 14 site_b 2021-07-01 CO2  8.5694332      14   0.05257841
#> 15 site_a 2021-08-01 CO2  8.1051011      15   0.08165216
#> 16 site_b 2021-08-01 CO2  6.8545510      16   0.03590167
#> 17 site_a 2021-09-01 CO2  6.2131739      17   0.05900213
#> 18 site_b 2021-09-01 CO2  5.1480204      18   0.04868803
#> 19 site_a 2021-10-01 CO2  4.1011940      19   0.09327191
#> 20 site_b 2021-10-01 CO2  3.2164744      20   0.06264923
#> 21 site_a 2021-11-01 CO2  1.6770558      21   0.09037221
#> 22 site_b 2021-11-01 CO2  1.5688890      22   0.03856545
#> 23 site_a 2021-12-01 CO2  1.1152357      23   0.07114110
#> 24 site_b 2021-12-01 CO2  0.9891861      24   0.02978153