Model runs as result of parameter variation

S3 class for the result of a call to alternate(). Essentially a list with added run_map and further attributes.

Usage

cfp_altres(
  x,
  og_model,
  f,
  run_map,
  return_raw = TRUE,
  error_funs = NULL,
  error_args = NULL
)

Arguments

x

A named list of cfp_pfres or cfp_fgres models.

og_model

The original model that was altered

f

A function taking in a soilphys object and recalculates the relevant columns. See complete_soilphys().

run_map

A data.frame created by run_map() with the necessary information how the data is to be changed with each distinct run_id.

return_raw

Should the models be returned as is, or after applying any error_funs. Default is TRUE - exporting the models.

error_funs

A list of functions to be applied after flux calculation if return_raw == FALSE. This can be used to output not the models but quality parameters instead. Output must contain the column RMSE.

error_args

A list of additional function arguments to be passed to any of the error_funs. Must match the length of error_funs

See also