Soil gas concentration data

Create a cfp_gasdata object. This is a data.frame containing gas concentration data for one or multiple soil profiles. Each soil profile is uniquely identified by columns in the data.frame specified by the id_cols attribute.

Usage

cfp_gasdata(x, ...)

# S3 method for class 'data.frame'
cfp_gasdata(x, id_cols, ...)

# S3 method for class 'cfp_dat'
cfp_gasdata(x, ...)

Arguments

x

A data.frame with the following columns:

gas

The gas of that observation.

depth (cm)

The depth of the observation.

x_ppm (ppm)

The concentration in ppm.

any of id_cols

All id_cols that identify one profile uniquely.

...

not used

id_cols

Column names in data.frame that uniquely identify each profile.

Value

A cfp_gasdata object.

See also

Examples

cfp_gasdata(
  ConFluxPro::gasdata,
  id_cols = c("site", "Date"))
#> 
#> added 'gas' to id_cols
#> 
#> A cfp_gasdata object 
#> id_cols: site Date gas 
#> 24  unique profiles 
#> 
#>       site       Date depth     x_ppm gas
#> 1   site_a 2021-01-01     5  422.7419 CO2
#> 2   site_a 2021-01-01     0  552.1412 CO2
#> 3   site_a 2021-01-01     0  555.8525 CO2
#> 4   site_a 2021-01-01     0  556.9315 CO2
#> 5   site_a 2021-01-01   -10  716.4256 CO2
#> 6   site_a 2021-01-01   -10  713.3633 CO2
#> 7   site_a 2021-01-01   -10  723.0691 CO2
#> 8   site_a 2021-01-01   -20  797.0534 CO2
#> 9   site_a 2021-01-01   -20  818.1842 CO2
#> 10  site_a 2021-01-01   -20  797.3729 CO2
#> 11  site_a 2021-01-01  -100 1111.5730 CO2
#> 12  site_a 2021-01-01  -100 1126.2997 CO2
#> 13  site_a 2021-01-01  -100 1071.1671 CO2
#> 14  site_b 2021-01-01     7  419.4424 CO2
#> 15  site_b 2021-01-01     0  566.4667 CO2
#> 16  site_b 2021-01-01     0  569.5438 CO2
#> 17  site_b 2021-01-01     0  565.8630 CO2
#> 18  site_b 2021-01-01   -10  656.0613 CO2
#> 19  site_b 2021-01-01   -10  657.3572 CO2
#> 20  site_b 2021-01-01   -10  679.9207 CO2
#> 21  site_b 2021-01-01   -20  794.9336 CO2
#> 22  site_b 2021-01-01   -20  780.1869 CO2
#> 23  site_b 2021-01-01   -20  796.2808 CO2
#> 24  site_b 2021-01-01  -100 1068.2201 CO2
#> 25  site_b 2021-01-01  -100 1078.4279 CO2
#> 26  site_b 2021-01-01  -100 1043.5430 CO2
#> 27  site_a 2021-02-01     5  419.2282 CO2
#> 28  site_a 2021-02-01     0  611.0210 CO2
#> 29  site_a 2021-02-01     0  624.3606 CO2
#> 30  site_a 2021-02-01     0  617.7600 CO2
#> 31  site_a 2021-02-01   -10  865.0991 CO2
#> 32  site_a 2021-02-01   -10  867.3435 CO2
#> 33  site_a 2021-02-01   -10  870.3159 CO2
#> 34  site_a 2021-02-01   -20  977.8661 CO2
#> 35  site_a 2021-02-01   -20  994.5743 CO2
#> 36  site_a 2021-02-01   -20  961.2449 CO2
#> 37  site_a 2021-02-01  -100 1410.3497 CO2
#> 38  site_a 2021-02-01  -100 1408.8546 CO2
#> 39  site_a 2021-02-01  -100 1373.6803 CO2
#> 40  site_b 2021-02-01     7  420.1084 CO2
#> 41  site_b 2021-02-01     0  641.7360 CO2
#> 42  site_b 2021-02-01     0  638.3337 CO2
#> 43  site_b 2021-02-01     0  645.0490 CO2
#> 44  site_b 2021-02-01   -10  791.4597 CO2
#> 45  site_b 2021-02-01   -10  785.6855 CO2
#> 46  site_b 2021-02-01   -10  801.8954 CO2
#> 47  site_b 2021-02-01   -20  974.8291 CO2
#> 48  site_b 2021-02-01   -20 1008.6615 CO2
#> 49  site_b 2021-02-01   -20  980.5283 CO2
#> 50  site_b 2021-02-01  -100 1379.7521 CO2
#> 51  site_b 2021-02-01  -100 1372.2433 CO2
#> 52  site_b 2021-02-01  -100 1347.3636 CO2
#> 53  site_a 2021-03-01     5  426.3029 CO2
#> 54  site_a 2021-03-01     0  693.9432 CO2
#> 55  site_a 2021-03-01     0  689.5181 CO2
#> 56  site_a 2021-03-01     0  691.0124 CO2
#> 57  site_a 2021-03-01   -10 1016.1515 CO2
#> 58  site_a 2021-03-01   -10 1009.0780 CO2
#> 59  site_a 2021-03-01   -10  972.0829 CO2
#> 60  site_a 2021-03-01   -20 1181.6977 CO2
#> 61  site_a 2021-03-01   -20 1168.6553 CO2
#> 62  site_a 2021-03-01   -20 1179.7046 CO2
#> 63  site_a 2021-03-01  -100 1781.4547 CO2
#> 64  site_a 2021-03-01  -100 1805.9921 CO2
#> 65  site_a 2021-03-01  -100 1742.1812 CO2
#> 66  site_b 2021-03-01     7  425.2102 CO2
#> 67  site_b 2021-03-01     0  716.6868 CO2
#> 68  site_b 2021-03-01     0  722.3080 CO2
#> 69  site_b 2021-03-01     0  721.3658 CO2
#> 70  site_b 2021-03-01   -10  932.6505 CO2
#> 71  site_b 2021-03-01   -10  911.4826 CO2
#> 72  site_b 2021-03-01   -10  922.9178 CO2
#> 73  site_b 2021-03-01   -20 1188.4704 CO2
#> 74  site_b 2021-03-01   -20 1156.9295 CO2
#> 75  site_b 2021-03-01   -20 1165.1434 CO2
#> 76  site_b 2021-03-01  -100 1697.4299 CO2
#> 77  site_b 2021-03-01  -100 1703.0454 CO2
#> 78  site_b 2021-03-01  -100 1693.9130 CO2
#> 79  site_a 2021-04-01     5  412.0280 CO2
#> 80  site_a 2021-04-01     0 1005.0039 CO2
#> 81  site_a 2021-04-01     0 1052.0287 CO2
#> 82  site_a 2021-04-01     0 1029.4426 CO2
#> 83  site_a 2021-04-01   -10 1745.3879 CO2
#> 84  site_a 2021-04-01   -10 1739.7358 CO2
#> 85  site_a 2021-04-01   -10 1710.7531 CO2
#> 86  site_a 2021-04-01   -20 2148.5399 CO2
#> 87  site_a 2021-04-01   -20 2111.2287 CO2
#> 88  site_a 2021-04-01   -20 2112.7759 CO2
#> 89  site_a 2021-04-01  -100 3507.0009 CO2
#> 90  site_a 2021-04-01  -100 3499.4697 CO2
#> 91  site_a 2021-04-01  -100 3537.9679 CO2
#> 92  site_b 2021-04-01     7  415.7144 CO2
#> 93  site_b 2021-04-01     0 1093.2063 CO2
#> 94  site_b 2021-04-01     0 1106.5400 CO2
#> 95  site_b 2021-04-01     0 1061.5058 CO2
#> 96  site_b 2021-04-01   -10 1530.7586 CO2
#> 97  site_b 2021-04-01   -10 1523.4431 CO2
#> 98  site_b 2021-04-01   -10 1514.6012 CO2
#> 99  site_b 2021-04-01   -20 2124.5992 CO2
#> 100 site_b 2021-04-01   -20 2150.3942 CO2
#> 101 site_b 2021-04-01   -20 2175.0434 CO2
#> 102 site_b 2021-04-01  -100 3325.5207 CO2
#> 103 site_b 2021-04-01  -100 3187.2834 CO2
#> 104 site_b 2021-04-01  -100 3379.7725 CO2
#> 105 site_a 2021-05-01     5  411.9987 CO2
#> 106 site_a 2021-05-01     0 1228.9323 CO2
#> 107 site_a 2021-05-01     0 1216.2659 CO2
#> 108 site_a 2021-05-01     0 1223.4636 CO2
#> 109 site_a 2021-05-01   -10 2190.7749 CO2
#> 110 site_a 2021-05-01   -10 2188.2898 CO2
#> 111 site_a 2021-05-01   -10 2183.0967 CO2
#> 112 site_a 2021-05-01   -20 2694.4844 CO2
#> 113 site_a 2021-05-01   -20 2658.8739 CO2
#> 114 site_a 2021-05-01   -20 2657.7470 CO2
#> 115 site_a 2021-05-01  -100 4302.5011 CO2
#> 116 site_a 2021-05-01  -100 4417.5900 CO2
#> 117 site_a 2021-05-01  -100 4405.8615 CO2
#> 118 site_b 2021-05-01     7  452.4227 CO2
#> 119 site_b 2021-05-01     0 1269.3092 CO2
#> 120 site_b 2021-05-01     0 1305.2941 CO2
#> 121 site_b 2021-05-01     0 1266.1530 CO2
#> 122 site_b 2021-05-01   -10 1879.0643 CO2
#> 123 site_b 2021-05-01   -10 1936.4893 CO2
#> 124 site_b 2021-05-01   -10 1896.1210 CO2
#> 125 site_b 2021-05-01   -20 2687.8906 CO2
#> 126 site_b 2021-05-01   -20 2662.3045 CO2
#> 127 site_b 2021-05-01   -20 2651.1797 CO2
#> 128 site_b 2021-05-01  -100 4017.7790 CO2
#> 129 site_b 2021-05-01  -100 4089.7727 CO2
#> 130 site_b 2021-05-01  -100 4216.1565 CO2
#> 131 site_a 2021-06-01     5  427.9467 CO2
#> 132 site_a 2021-06-01     0 1346.9994 CO2
#> 133 site_a 2021-06-01     0 1360.8018 CO2
#> 134 site_a 2021-06-01     0 1392.2409 CO2
#> 135 site_a 2021-06-01   -10 2538.4739 CO2
#> 136 site_a 2021-06-01   -10 2431.9212 CO2
#> 137 site_a 2021-06-01   -10 2473.0726 CO2
#> 138 site_a 2021-06-01   -20 3150.1049 CO2
#> 139 site_a 2021-06-01   -20 3033.1189 CO2
#> 140 site_a 2021-06-01   -20 3062.3271 CO2
#> 141 site_a 2021-06-01  -100 5114.9587 CO2
#> 142 site_a 2021-06-01  -100 5030.7937 CO2
#> 143 site_a 2021-06-01  -100 5077.3082 CO2
#> 144 site_b 2021-06-01     7  419.5878 CO2
#> 145 site_b 2021-06-01     0 1437.4117 CO2
#> 146 site_b 2021-06-01     0 1480.1748 CO2
#> 147 site_b 2021-06-01     0 1435.5322 CO2
#> 148 site_b 2021-06-01   -10 2165.8069 CO2
#> 149 site_b 2021-06-01   -10 2213.2682 CO2
#> 150 site_b 2021-06-01   -10 2139.6325 CO2
#> 151 site_b 2021-06-01   -20 3063.1511 CO2
#> 152 site_b 2021-06-01   -20 2957.3919 CO2
#> 153 site_b 2021-06-01   -20 3147.7019 CO2
#> 154 site_b 2021-06-01  -100 4801.2672 CO2
#> 155 site_b 2021-06-01  -100 4780.8762 CO2
#> 156 site_b 2021-06-01  -100 4699.9835 CO2
#> 157 site_a 2021-07-01     5  419.8448 CO2
#> 158 site_a 2021-07-01     0 1731.9887 CO2
#> 159 site_a 2021-07-01     0 1742.6603 CO2
#> 160 site_a 2021-07-01     0 1815.5070 CO2
#> 161 site_a 2021-07-01   -10 3349.4684 CO2
#> 162 site_a 2021-07-01   -10 3295.6931 CO2
#> 163 site_a 2021-07-01   -10 3369.7924 CO2
#> 164 site_a 2021-07-01   -20 4163.7262 CO2
#> 165 site_a 2021-07-01   -20 4259.0014 CO2
#> 166 site_a 2021-07-01   -20 4343.2422 CO2
#> 167 site_a 2021-07-01  -100 6991.0961 CO2
#> 168 site_a 2021-07-01  -100 7208.1975 CO2
#> 169 site_a 2021-07-01  -100 7152.7347 CO2
#> 170 site_b 2021-07-01     7  437.9113 CO2
#> 171 site_b 2021-07-01     0 1880.8089 CO2
#> 172 site_b 2021-07-01     0 1923.4648 CO2
#> 173 site_b 2021-07-01     0 1820.1978 CO2
#> 174 site_b 2021-07-01   -10 3041.3675 CO2
#> 175 site_b 2021-07-01   -10 2991.8867 CO2
#> 176 site_b 2021-07-01   -10 2930.9534 CO2
#> 177 site_b 2021-07-01   -20 4055.0993 CO2
#> 178 site_b 2021-07-01   -20 4264.3009 CO2
#> 179 site_b 2021-07-01   -20 4248.3194 CO2
#> 180 site_b 2021-07-01  -100 6719.0467 CO2
#> 181 site_b 2021-07-01  -100 6742.7184 CO2
#> 182 site_b 2021-07-01  -100 6632.3837 CO2
#> 183 site_a 2021-08-01     5  425.9009 CO2
#> 184 site_a 2021-08-01     0 1504.6289 CO2
#> 185 site_a 2021-08-01     0 1486.2637 CO2
#> 186 site_a 2021-08-01     0 1452.4404 CO2
#> 187 site_a 2021-08-01   -10 2805.6359 CO2
#> 188 site_a 2021-08-01   -10 2798.6014 CO2
#> 189 site_a 2021-08-01   -10 2731.8573 CO2
#> 190 site_a 2021-08-01   -20 3316.4329 CO2
#> 191 site_a 2021-08-01   -20 3460.3967 CO2
#> 192 site_a 2021-08-01   -20 3416.3930 CO2
#> 193 site_a 2021-08-01  -100 5826.3134 CO2
#> 194 site_a 2021-08-01  -100 5640.7197 CO2
#> 195 site_a 2021-08-01  -100 5680.8995 CO2
#> 196 site_b 2021-08-01     7  437.3724 CO2
#> 197 site_b 2021-08-01     0 1608.9208 CO2
#> 198 site_b 2021-08-01     0 1614.7676 CO2
#> 199 site_b 2021-08-01     0 1654.0873 CO2
#> 200 site_b 2021-08-01   -10 2442.1834 CO2
#> 201 site_b 2021-08-01   -10 2339.9554 CO2
#> 202 site_b 2021-08-01   -10 2452.0213 CO2
#> 203 site_b 2021-08-01   -20 3537.7060 CO2
#> 204 site_b 2021-08-01   -20 3608.7631 CO2
#> 205 site_b 2021-08-01   -20 3333.8511 CO2
#> 206 site_b 2021-08-01  -100 5321.8973 CO2
#> 207 site_b 2021-08-01  -100 5375.3014 CO2
#> 208 site_b 2021-08-01  -100 5333.5133 CO2
#> 209 site_a 2021-09-01     5  412.2511 CO2
#> 210 site_a 2021-09-01     0 1221.9509 CO2
#> 211 site_a 2021-09-01     0 1197.5707 CO2
#> 212 site_a 2021-09-01     0 1275.2873 CO2
#> 213 site_a 2021-09-01   -10 2187.8799 CO2
#> 214 site_a 2021-09-01   -10 2180.9721 CO2
#> 215 site_a 2021-09-01   -10 2201.8423 CO2
#> 216 site_a 2021-09-01   -20 2690.2449 CO2
#> 217 site_a 2021-09-01   -20 2680.0747 CO2
#> 218 site_a 2021-09-01   -20 2776.6072 CO2
#> 219 site_a 2021-09-01  -100 4432.4673 CO2
#> 220 site_a 2021-09-01  -100 4336.4758 CO2
#> 221 site_a 2021-09-01  -100 4486.7101 CO2
#> 222 site_b 2021-09-01     7  415.7818 CO2
#> 223 site_b 2021-09-01     0 1289.4769 CO2
#> 224 site_b 2021-09-01     0 1276.3649 CO2
#> 225 site_b 2021-09-01     0 1312.2808 CO2
#> 226 site_b 2021-09-01   -10 1925.7353 CO2
#> 227 site_b 2021-09-01   -10 1950.5640 CO2
#> 228 site_b 2021-09-01   -10 1923.5629 CO2
#> 229 site_b 2021-09-01   -20 2648.4898 CO2
#> 230 site_b 2021-09-01   -20 2763.0142 CO2
#> 231 site_b 2021-09-01   -20 2671.6942 CO2
#> 232 site_b 2021-09-01  -100 4285.3157 CO2
#> 233 site_b 2021-09-01  -100 4091.8576 CO2
#> 234 site_b 2021-09-01  -100 4158.0496 CO2
#> 235 site_a 2021-10-01     5  417.8452 CO2
#> 236 site_a 2021-10-01     0  951.3446 CO2
#> 237 site_a 2021-10-01     0  979.1793 CO2
#> 238 site_a 2021-10-01     0  957.2358 CO2
#> 239 site_a 2021-10-01   -10 1601.8614 CO2
#> 240 site_a 2021-10-01   -10 1573.3831 CO2
#> 241 site_a 2021-10-01   -10 1578.4988 CO2
#> 242 site_a 2021-10-01   -20 1971.9228 CO2
#> 243 site_a 2021-10-01   -20 1982.3393 CO2
#> 244 site_a 2021-10-01   -20 1981.9545 CO2
#> 245 site_a 2021-10-01  -100 3025.1618 CO2
#> 246 site_a 2021-10-01  -100 3230.9976 CO2
#> 247 site_a 2021-10-01  -100 3169.0124 CO2
#> 248 site_b 2021-10-01     7  419.7863 CO2
#> 249 site_b 2021-10-01     0 1019.2577 CO2
#> 250 site_b 2021-10-01     0  992.4578 CO2
#> 251 site_b 2021-10-01     0  990.4615 CO2
#> 252 site_b 2021-10-01   -10 1429.1772 CO2
#> 253 site_b 2021-10-01   -10 1399.2361 CO2
#> 254 site_b 2021-10-01   -10 1432.5605 CO2
#> 255 site_b 2021-10-01   -20 1983.9082 CO2
#> 256 site_b 2021-10-01   -20 1890.6451 CO2
#> 257 site_b 2021-10-01   -20 1902.4624 CO2
#> 258 site_b 2021-10-01  -100 2942.7938 CO2
#> 259 site_b 2021-10-01  -100 2878.9442 CO2
#> 260 site_b 2021-10-01  -100 2917.0030 CO2
#> 261 site_a 2021-11-01     5  423.4910 CO2
#> 262 site_a 2021-11-01     0  696.5564 CO2
#> 263 site_a 2021-11-01     0  691.8708 CO2
#> 264 site_a 2021-11-01     0  673.9876 CO2
#> 265 site_a 2021-11-01   -10 1007.3520 CO2
#> 266 site_a 2021-11-01   -10 1020.7773 CO2
#> 267 site_a 2021-11-01   -10 1017.7583 CO2
#> 268 site_a 2021-11-01   -20 1172.1837 CO2
#> 269 site_a 2021-11-01   -20 1122.0014 CO2
#> 270 site_a 2021-11-01   -20 1177.2193 CO2
#> 271 site_a 2021-11-01  -100 1781.2126 CO2
#> 272 site_a 2021-11-01  -100 1765.3741 CO2
#> 273 site_a 2021-11-01  -100 1768.7224 CO2
#> 274 site_b 2021-11-01     7  421.7263 CO2
#> 275 site_b 2021-11-01     0  710.8276 CO2
#> 276 site_b 2021-11-01     0  724.4798 CO2
#> 277 site_b 2021-11-01     0  717.7631 CO2
#> 278 site_b 2021-11-01   -10  909.0880 CO2
#> 279 site_b 2021-11-01   -10  940.5789 CO2
#> 280 site_b 2021-11-01   -10  938.4535 CO2
#> 281 site_b 2021-11-01   -20 1192.4815 CO2
#> 282 site_b 2021-11-01   -20 1142.7474 CO2
#> 283 site_b 2021-11-01   -20 1164.6139 CO2
#> 284 site_b 2021-11-01  -100 1699.0654 CO2
#> 285 site_b 2021-11-01  -100 1681.3117 CO2
#> 286 site_b 2021-11-01  -100 1690.4404 CO2
#> 287 site_a 2021-12-01     5  427.3788 CO2
#> 288 site_a 2021-12-01     0  616.6897 CO2
#> 289 site_a 2021-12-01     0  608.9208 CO2
#> 290 site_a 2021-12-01     0  623.2422 CO2
#> 291 site_a 2021-12-01   -10  854.8116 CO2
#> 292 site_a 2021-12-01   -10  865.0144 CO2
#> 293 site_a 2021-12-01   -10  853.6885 CO2
#> 294 site_a 2021-12-01   -20 1020.1808 CO2
#> 295 site_a 2021-12-01   -20  985.1444 CO2
#> 296 site_a 2021-12-01   -20  979.8400 CO2
#> 297 site_a 2021-12-01  -100 1424.2591 CO2
#> 298 site_a 2021-12-01  -100 1447.4077 CO2
#> 299 site_a 2021-12-01  -100 1430.1902 CO2
#> 300 site_b 2021-12-01     7  415.1231 CO2
#> 301 site_b 2021-12-01     0  640.4723 CO2
#> 302 site_b 2021-12-01     0  645.0615 CO2
#> 303 site_b 2021-12-01     0  640.7339 CO2
#> 304 site_b 2021-12-01   -10  804.6156 CO2
#> 305 site_b 2021-12-01   -10  796.6817 CO2
#> 306 site_b 2021-12-01   -10  797.4790 CO2
#> 307 site_b 2021-12-01   -20  994.2355 CO2
#> 308 site_b 2021-12-01   -20 1001.8942 CO2
#> 309 site_b 2021-12-01   -20  968.3041 CO2
#> 310 site_b 2021-12-01  -100 1364.2465 CO2
#> 311 site_b 2021-12-01  -100 1363.4034 CO2
#> 312 site_b 2021-12-01  -100 1347.9243 CO2
### Also used to extract the gasdata object from cfp_dat
cfp_gasdata(ConFluxPro::base_dat)
#> 
#> A cfp_gasdata object 
#> id_cols: site Date gas 
#> 24  unique profiles 
#> 
#>       site       Date gas depth     x_ppm
#> 1   site_a 2021-01-01 CO2     5  422.7419
#> 2   site_a 2021-01-01 CO2     0  552.1412
#> 3   site_a 2021-01-01 CO2     0  555.8525
#> 4   site_a 2021-01-01 CO2     0  556.9315
#> 5   site_a 2021-01-01 CO2   -10  716.4256
#> 6   site_a 2021-01-01 CO2   -10  713.3633
#> 7   site_a 2021-01-01 CO2   -10  723.0691
#> 8   site_a 2021-01-01 CO2   -20  797.0534
#> 9   site_a 2021-01-01 CO2   -20  818.1842
#> 10  site_a 2021-01-01 CO2   -20  797.3729
#> 11  site_a 2021-01-01 CO2  -100 1111.5730
#> 12  site_a 2021-01-01 CO2  -100 1126.2997
#> 13  site_a 2021-01-01 CO2  -100 1071.1671
#> 14  site_b 2021-01-01 CO2     7  419.4424
#> 15  site_b 2021-01-01 CO2     0  566.4667
#> 16  site_b 2021-01-01 CO2     0  569.5438
#> 17  site_b 2021-01-01 CO2     0  565.8630
#> 18  site_b 2021-01-01 CO2   -10  656.0613
#> 19  site_b 2021-01-01 CO2   -10  657.3572
#> 20  site_b 2021-01-01 CO2   -10  679.9207
#> 21  site_b 2021-01-01 CO2   -20  794.9336
#> 22  site_b 2021-01-01 CO2   -20  780.1869
#> 23  site_b 2021-01-01 CO2   -20  796.2808
#> 24  site_b 2021-01-01 CO2  -100 1068.2201
#> 25  site_b 2021-01-01 CO2  -100 1078.4279
#> 26  site_b 2021-01-01 CO2  -100 1043.5430
#> 27  site_a 2021-02-01 CO2     5  419.2282
#> 28  site_a 2021-02-01 CO2     0  611.0210
#> 29  site_a 2021-02-01 CO2     0  624.3606
#> 30  site_a 2021-02-01 CO2     0  617.7600
#> 31  site_a 2021-02-01 CO2   -10  865.0991
#> 32  site_a 2021-02-01 CO2   -10  867.3435
#> 33  site_a 2021-02-01 CO2   -10  870.3159
#> 34  site_a 2021-02-01 CO2   -20  977.8661
#> 35  site_a 2021-02-01 CO2   -20  994.5743
#> 36  site_a 2021-02-01 CO2   -20  961.2449
#> 37  site_a 2021-02-01 CO2  -100 1410.3497
#> 38  site_a 2021-02-01 CO2  -100 1408.8546
#> 39  site_a 2021-02-01 CO2  -100 1373.6803
#> 40  site_b 2021-02-01 CO2     7  420.1084
#> 41  site_b 2021-02-01 CO2     0  641.7360
#> 42  site_b 2021-02-01 CO2     0  638.3337
#> 43  site_b 2021-02-01 CO2     0  645.0490
#> 44  site_b 2021-02-01 CO2   -10  791.4597
#> 45  site_b 2021-02-01 CO2   -10  785.6855
#> 46  site_b 2021-02-01 CO2   -10  801.8954
#> 47  site_b 2021-02-01 CO2   -20  974.8291
#> 48  site_b 2021-02-01 CO2   -20 1008.6615
#> 49  site_b 2021-02-01 CO2   -20  980.5283
#> 50  site_b 2021-02-01 CO2  -100 1379.7521
#> 51  site_b 2021-02-01 CO2  -100 1372.2433
#> 52  site_b 2021-02-01 CO2  -100 1347.3636
#> 53  site_a 2021-03-01 CO2     5  426.3029
#> 54  site_a 2021-03-01 CO2     0  693.9432
#> 55  site_a 2021-03-01 CO2     0  689.5181
#> 56  site_a 2021-03-01 CO2     0  691.0124
#> 57  site_a 2021-03-01 CO2   -10 1016.1515
#> 58  site_a 2021-03-01 CO2   -10 1009.0780
#> 59  site_a 2021-03-01 CO2   -10  972.0829
#> 60  site_a 2021-03-01 CO2   -20 1181.6977
#> 61  site_a 2021-03-01 CO2   -20 1168.6553
#> 62  site_a 2021-03-01 CO2   -20 1179.7046
#> 63  site_a 2021-03-01 CO2  -100 1781.4547
#> 64  site_a 2021-03-01 CO2  -100 1805.9921
#> 65  site_a 2021-03-01 CO2  -100 1742.1812
#> 66  site_b 2021-03-01 CO2     7  425.2102
#> 67  site_b 2021-03-01 CO2     0  716.6868
#> 68  site_b 2021-03-01 CO2     0  722.3080
#> 69  site_b 2021-03-01 CO2     0  721.3658
#> 70  site_b 2021-03-01 CO2   -10  932.6505
#> 71  site_b 2021-03-01 CO2   -10  911.4826
#> 72  site_b 2021-03-01 CO2   -10  922.9178
#> 73  site_b 2021-03-01 CO2   -20 1188.4704
#> 74  site_b 2021-03-01 CO2   -20 1156.9295
#> 75  site_b 2021-03-01 CO2   -20 1165.1434
#> 76  site_b 2021-03-01 CO2  -100 1697.4299
#> 77  site_b 2021-03-01 CO2  -100 1703.0454
#> 78  site_b 2021-03-01 CO2  -100 1693.9130
#> 79  site_a 2021-04-01 CO2     5  412.0280
#> 80  site_a 2021-04-01 CO2     0 1005.0039
#> 81  site_a 2021-04-01 CO2     0 1052.0287
#> 82  site_a 2021-04-01 CO2     0 1029.4426
#> 83  site_a 2021-04-01 CO2   -10 1745.3879
#> 84  site_a 2021-04-01 CO2   -10 1739.7358
#> 85  site_a 2021-04-01 CO2   -10 1710.7531
#> 86  site_a 2021-04-01 CO2   -20 2148.5399
#> 87  site_a 2021-04-01 CO2   -20 2111.2287
#> 88  site_a 2021-04-01 CO2   -20 2112.7759
#> 89  site_a 2021-04-01 CO2  -100 3507.0009
#> 90  site_a 2021-04-01 CO2  -100 3499.4697
#> 91  site_a 2021-04-01 CO2  -100 3537.9679
#> 92  site_b 2021-04-01 CO2     7  415.7144
#> 93  site_b 2021-04-01 CO2     0 1093.2063
#> 94  site_b 2021-04-01 CO2     0 1106.5400
#> 95  site_b 2021-04-01 CO2     0 1061.5058
#> 96  site_b 2021-04-01 CO2   -10 1530.7586
#> 97  site_b 2021-04-01 CO2   -10 1523.4431
#> 98  site_b 2021-04-01 CO2   -10 1514.6012
#> 99  site_b 2021-04-01 CO2   -20 2124.5992
#> 100 site_b 2021-04-01 CO2   -20 2150.3942
#> 101 site_b 2021-04-01 CO2   -20 2175.0434
#> 102 site_b 2021-04-01 CO2  -100 3325.5207
#> 103 site_b 2021-04-01 CO2  -100 3187.2834
#> 104 site_b 2021-04-01 CO2  -100 3379.7725
#> 105 site_a 2021-05-01 CO2     5  411.9987
#> 106 site_a 2021-05-01 CO2     0 1228.9323
#> 107 site_a 2021-05-01 CO2     0 1216.2659
#> 108 site_a 2021-05-01 CO2     0 1223.4636
#> 109 site_a 2021-05-01 CO2   -10 2190.7749
#> 110 site_a 2021-05-01 CO2   -10 2188.2898
#> 111 site_a 2021-05-01 CO2   -10 2183.0967
#> 112 site_a 2021-05-01 CO2   -20 2694.4844
#> 113 site_a 2021-05-01 CO2   -20 2658.8739
#> 114 site_a 2021-05-01 CO2   -20 2657.7470
#> 115 site_a 2021-05-01 CO2  -100 4302.5011
#> 116 site_a 2021-05-01 CO2  -100 4417.5900
#> 117 site_a 2021-05-01 CO2  -100 4405.8615
#> 118 site_b 2021-05-01 CO2     7  452.4227
#> 119 site_b 2021-05-01 CO2     0 1269.3092
#> 120 site_b 2021-05-01 CO2     0 1305.2941
#> 121 site_b 2021-05-01 CO2     0 1266.1530
#> 122 site_b 2021-05-01 CO2   -10 1879.0643
#> 123 site_b 2021-05-01 CO2   -10 1936.4893
#> 124 site_b 2021-05-01 CO2   -10 1896.1210
#> 125 site_b 2021-05-01 CO2   -20 2687.8906
#> 126 site_b 2021-05-01 CO2   -20 2662.3045
#> 127 site_b 2021-05-01 CO2   -20 2651.1797
#> 128 site_b 2021-05-01 CO2  -100 4017.7790
#> 129 site_b 2021-05-01 CO2  -100 4089.7727
#> 130 site_b 2021-05-01 CO2  -100 4216.1565
#> 131 site_a 2021-06-01 CO2     5  427.9467
#> 132 site_a 2021-06-01 CO2     0 1346.9994
#> 133 site_a 2021-06-01 CO2     0 1360.8018
#> 134 site_a 2021-06-01 CO2     0 1392.2409
#> 135 site_a 2021-06-01 CO2   -10 2538.4739
#> 136 site_a 2021-06-01 CO2   -10 2431.9212
#> 137 site_a 2021-06-01 CO2   -10 2473.0726
#> 138 site_a 2021-06-01 CO2   -20 3150.1049
#> 139 site_a 2021-06-01 CO2   -20 3033.1189
#> 140 site_a 2021-06-01 CO2   -20 3062.3271
#> 141 site_a 2021-06-01 CO2  -100 5114.9587
#> 142 site_a 2021-06-01 CO2  -100 5030.7937
#> 143 site_a 2021-06-01 CO2  -100 5077.3082
#> 144 site_b 2021-06-01 CO2     7  419.5878
#> 145 site_b 2021-06-01 CO2     0 1437.4117
#> 146 site_b 2021-06-01 CO2     0 1480.1748
#> 147 site_b 2021-06-01 CO2     0 1435.5322
#> 148 site_b 2021-06-01 CO2   -10 2165.8069
#> 149 site_b 2021-06-01 CO2   -10 2213.2682
#> 150 site_b 2021-06-01 CO2   -10 2139.6325
#> 151 site_b 2021-06-01 CO2   -20 3063.1511
#> 152 site_b 2021-06-01 CO2   -20 2957.3919
#> 153 site_b 2021-06-01 CO2   -20 3147.7019
#> 154 site_b 2021-06-01 CO2  -100 4801.2672
#> 155 site_b 2021-06-01 CO2  -100 4780.8762
#> 156 site_b 2021-06-01 CO2  -100 4699.9835
#> 157 site_a 2021-07-01 CO2     5  419.8448
#> 158 site_a 2021-07-01 CO2     0 1731.9887
#> 159 site_a 2021-07-01 CO2     0 1742.6603
#> 160 site_a 2021-07-01 CO2     0 1815.5070
#> 161 site_a 2021-07-01 CO2   -10 3349.4684
#> 162 site_a 2021-07-01 CO2   -10 3295.6931
#> 163 site_a 2021-07-01 CO2   -10 3369.7924
#> 164 site_a 2021-07-01 CO2   -20 4163.7262
#> 165 site_a 2021-07-01 CO2   -20 4259.0014
#> 166 site_a 2021-07-01 CO2   -20 4343.2422
#> 167 site_a 2021-07-01 CO2  -100 6991.0961
#> 168 site_a 2021-07-01 CO2  -100 7208.1975
#> 169 site_a 2021-07-01 CO2  -100 7152.7347
#> 170 site_b 2021-07-01 CO2     7  437.9113
#> 171 site_b 2021-07-01 CO2     0 1880.8089
#> 172 site_b 2021-07-01 CO2     0 1923.4648
#> 173 site_b 2021-07-01 CO2     0 1820.1978
#> 174 site_b 2021-07-01 CO2   -10 3041.3675
#> 175 site_b 2021-07-01 CO2   -10 2991.8867
#> 176 site_b 2021-07-01 CO2   -10 2930.9534
#> 177 site_b 2021-07-01 CO2   -20 4055.0993
#> 178 site_b 2021-07-01 CO2   -20 4264.3009
#> 179 site_b 2021-07-01 CO2   -20 4248.3194
#> 180 site_b 2021-07-01 CO2  -100 6719.0467
#> 181 site_b 2021-07-01 CO2  -100 6742.7184
#> 182 site_b 2021-07-01 CO2  -100 6632.3837
#> 183 site_a 2021-08-01 CO2     5  425.9009
#> 184 site_a 2021-08-01 CO2     0 1504.6289
#> 185 site_a 2021-08-01 CO2     0 1486.2637
#> 186 site_a 2021-08-01 CO2     0 1452.4404
#> 187 site_a 2021-08-01 CO2   -10 2805.6359
#> 188 site_a 2021-08-01 CO2   -10 2798.6014
#> 189 site_a 2021-08-01 CO2   -10 2731.8573
#> 190 site_a 2021-08-01 CO2   -20 3316.4329
#> 191 site_a 2021-08-01 CO2   -20 3460.3967
#> 192 site_a 2021-08-01 CO2   -20 3416.3930
#> 193 site_a 2021-08-01 CO2  -100 5826.3134
#> 194 site_a 2021-08-01 CO2  -100 5640.7197
#> 195 site_a 2021-08-01 CO2  -100 5680.8995
#> 196 site_b 2021-08-01 CO2     7  437.3724
#> 197 site_b 2021-08-01 CO2     0 1608.9208
#> 198 site_b 2021-08-01 CO2     0 1614.7676
#> 199 site_b 2021-08-01 CO2     0 1654.0873
#> 200 site_b 2021-08-01 CO2   -10 2442.1834
#> 201 site_b 2021-08-01 CO2   -10 2339.9554
#> 202 site_b 2021-08-01 CO2   -10 2452.0213
#> 203 site_b 2021-08-01 CO2   -20 3537.7060
#> 204 site_b 2021-08-01 CO2   -20 3608.7631
#> 205 site_b 2021-08-01 CO2   -20 3333.8511
#> 206 site_b 2021-08-01 CO2  -100 5321.8973
#> 207 site_b 2021-08-01 CO2  -100 5375.3014
#> 208 site_b 2021-08-01 CO2  -100 5333.5133
#> 209 site_a 2021-09-01 CO2     5  412.2511
#> 210 site_a 2021-09-01 CO2     0 1221.9509
#> 211 site_a 2021-09-01 CO2     0 1197.5707
#> 212 site_a 2021-09-01 CO2     0 1275.2873
#> 213 site_a 2021-09-01 CO2   -10 2187.8799
#> 214 site_a 2021-09-01 CO2   -10 2180.9721
#> 215 site_a 2021-09-01 CO2   -10 2201.8423
#> 216 site_a 2021-09-01 CO2   -20 2690.2449
#> 217 site_a 2021-09-01 CO2   -20 2680.0747
#> 218 site_a 2021-09-01 CO2   -20 2776.6072
#> 219 site_a 2021-09-01 CO2  -100 4432.4673
#> 220 site_a 2021-09-01 CO2  -100 4336.4758
#> 221 site_a 2021-09-01 CO2  -100 4486.7101
#> 222 site_b 2021-09-01 CO2     7  415.7818
#> 223 site_b 2021-09-01 CO2     0 1289.4769
#> 224 site_b 2021-09-01 CO2     0 1276.3649
#> 225 site_b 2021-09-01 CO2     0 1312.2808
#> 226 site_b 2021-09-01 CO2   -10 1925.7353
#> 227 site_b 2021-09-01 CO2   -10 1950.5640
#> 228 site_b 2021-09-01 CO2   -10 1923.5629
#> 229 site_b 2021-09-01 CO2   -20 2648.4898
#> 230 site_b 2021-09-01 CO2   -20 2763.0142
#> 231 site_b 2021-09-01 CO2   -20 2671.6942
#> 232 site_b 2021-09-01 CO2  -100 4285.3157
#> 233 site_b 2021-09-01 CO2  -100 4091.8576
#> 234 site_b 2021-09-01 CO2  -100 4158.0496
#> 235 site_a 2021-10-01 CO2     5  417.8452
#> 236 site_a 2021-10-01 CO2     0  951.3446
#> 237 site_a 2021-10-01 CO2     0  979.1793
#> 238 site_a 2021-10-01 CO2     0  957.2358
#> 239 site_a 2021-10-01 CO2   -10 1601.8614
#> 240 site_a 2021-10-01 CO2   -10 1573.3831
#> 241 site_a 2021-10-01 CO2   -10 1578.4988
#> 242 site_a 2021-10-01 CO2   -20 1971.9228
#> 243 site_a 2021-10-01 CO2   -20 1982.3393
#> 244 site_a 2021-10-01 CO2   -20 1981.9545
#> 245 site_a 2021-10-01 CO2  -100 3025.1618
#> 246 site_a 2021-10-01 CO2  -100 3230.9976
#> 247 site_a 2021-10-01 CO2  -100 3169.0124
#> 248 site_b 2021-10-01 CO2     7  419.7863
#> 249 site_b 2021-10-01 CO2     0 1019.2577
#> 250 site_b 2021-10-01 CO2     0  992.4578
#> 251 site_b 2021-10-01 CO2     0  990.4615
#> 252 site_b 2021-10-01 CO2   -10 1429.1772
#> 253 site_b 2021-10-01 CO2   -10 1399.2361
#> 254 site_b 2021-10-01 CO2   -10 1432.5605
#> 255 site_b 2021-10-01 CO2   -20 1983.9082
#> 256 site_b 2021-10-01 CO2   -20 1890.6451
#> 257 site_b 2021-10-01 CO2   -20 1902.4624
#> 258 site_b 2021-10-01 CO2  -100 2942.7938
#> 259 site_b 2021-10-01 CO2  -100 2878.9442
#> 260 site_b 2021-10-01 CO2  -100 2917.0030
#> 261 site_a 2021-11-01 CO2     5  423.4910
#> 262 site_a 2021-11-01 CO2     0  696.5564
#> 263 site_a 2021-11-01 CO2     0  691.8708
#> 264 site_a 2021-11-01 CO2     0  673.9876
#> 265 site_a 2021-11-01 CO2   -10 1007.3520
#> 266 site_a 2021-11-01 CO2   -10 1020.7773
#> 267 site_a 2021-11-01 CO2   -10 1017.7583
#> 268 site_a 2021-11-01 CO2   -20 1172.1837
#> 269 site_a 2021-11-01 CO2   -20 1122.0014
#> 270 site_a 2021-11-01 CO2   -20 1177.2193
#> 271 site_a 2021-11-01 CO2  -100 1781.2126
#> 272 site_a 2021-11-01 CO2  -100 1765.3741
#> 273 site_a 2021-11-01 CO2  -100 1768.7224
#> 274 site_b 2021-11-01 CO2     7  421.7263
#> 275 site_b 2021-11-01 CO2     0  710.8276
#> 276 site_b 2021-11-01 CO2     0  724.4798
#> 277 site_b 2021-11-01 CO2     0  717.7631
#> 278 site_b 2021-11-01 CO2   -10  909.0880
#> 279 site_b 2021-11-01 CO2   -10  940.5789
#> 280 site_b 2021-11-01 CO2   -10  938.4535
#> 281 site_b 2021-11-01 CO2   -20 1192.4815
#> 282 site_b 2021-11-01 CO2   -20 1142.7474
#> 283 site_b 2021-11-01 CO2   -20 1164.6139
#> 284 site_b 2021-11-01 CO2  -100 1699.0654
#> 285 site_b 2021-11-01 CO2  -100 1681.3117
#> 286 site_b 2021-11-01 CO2  -100 1690.4404
#> 287 site_a 2021-12-01 CO2     5  427.3788
#> 288 site_a 2021-12-01 CO2     0  616.6897
#> 289 site_a 2021-12-01 CO2     0  608.9208
#> 290 site_a 2021-12-01 CO2     0  623.2422
#> 291 site_a 2021-12-01 CO2   -10  854.8116
#> 292 site_a 2021-12-01 CO2   -10  865.0144
#> 293 site_a 2021-12-01 CO2   -10  853.6885
#> 294 site_a 2021-12-01 CO2   -20 1020.1808
#> 295 site_a 2021-12-01 CO2   -20  985.1444
#> 296 site_a 2021-12-01 CO2   -20  979.8400
#> 297 site_a 2021-12-01 CO2  -100 1424.2591
#> 298 site_a 2021-12-01 CO2  -100 1447.4077
#> 299 site_a 2021-12-01 CO2  -100 1430.1902
#> 300 site_b 2021-12-01 CO2     7  415.1231
#> 301 site_b 2021-12-01 CO2     0  640.4723
#> 302 site_b 2021-12-01 CO2     0  645.0615
#> 303 site_b 2021-12-01 CO2     0  640.7339
#> 304 site_b 2021-12-01 CO2   -10  804.6156
#> 305 site_b 2021-12-01 CO2   -10  796.6817
#> 306 site_b 2021-12-01 CO2   -10  797.4790
#> 307 site_b 2021-12-01 CO2   -20  994.2355
#> 308 site_b 2021-12-01 CO2   -20 1001.8942
#> 309 site_b 2021-12-01 CO2   -20  968.3041
#> 310 site_b 2021-12-01 CO2  -100 1364.2465
#> 311 site_b 2021-12-01 CO2  -100 1363.4034
#> 312 site_b 2021-12-01 CO2  -100 1347.9243