Model frame for pro_flux

An S3 class for pro_flux() models. The class inherits from cfp_dat and adds any model specific parameters.

Usage

cfp_pfmod(
  x,
  zero_flux = TRUE,
  zero_limits = c(-Inf, Inf),
  DSD0_optim = FALSE,
  evenness_factor = 0,
  known_flux_factor = 0
)

Arguments

x

A cfp_dat object with all the necessary input datasets.

zero_flux

(logical) Applies the zero-flux boundary condition? If FALSE, F0 is optimized alongside the production rates.

zero_limits

(numeric vector) a vector of length 2 defining the lower and upper limit of the lowest flux if zero_flux = FALSE.

DSD0_optim

evenness_factor

(numeric) A user defined factor used to penalise strong differences between the optimised production rates. This must be identified by trial-and-error and can help prevent that production rates are simply set to zero basically the lower a production is relative to the the maximum of the absolute of all productions, the higher it is penalised. The evenness_factor then defines the weight of this penalty in the optimisation algorithm prod_optim.

known_flux_factor

Value

A cfp_pfmod object that inherits from cfp_dat()

See also

Examples

cfp_pfmod(ConFluxPro::base_dat)
#> 
#> A cfp_pfmod pro_flux model. 
#> zero_flux: TRUE  
#> zero_limits:  -Inf Inf  
#> DSD0_optim:  FALSE  
#> evenness_factor:  0  
#> known_flux_factor:  0  
#> 
#> A cfp_dat object to be used as input in ConFluxPro models. 
#> id_cols: site Date gas 
#> number of profiles:  24 
#> number of groups:  2 
#>