Model result of pro_flux

A function to create an object of class cfp_pfres. This is the central result class generated by running pro_flux(). Intended for internal use only.

Usage

cfp_pfres(x, y)

Arguments

x

A valid cfp_pfmod object

y

The corresponding PROFULX data.frame.

Value

A cfp_pfres object. This inherits from cfp_pfmod().

See also

Examples

if (FALSE) { # interactive()
PROFLUX <- pro_flux(ConFluxPro::base_dat)
cfp_pfres(
  cfp_pfmod(ConFluxPro::base_dat),
  PROFLUX$PROFLUX
)

}