Interpolate over depth to layered profile

Interpolate and discretize data into a layered structure. The output is a data.frame where each profile is separated into layers that intersect at depths defined in the function call. See cfp_layered_profile().

There are different interpolation methods implemented, which might be more practical for different parameters or tasks.

A 'linear' interpolation for continuous parameters, (e.g. soil temperature).

The 'boundary' interpolation is only suitable for data that is already layered. It selects the value from the old layer that in which the new layer will lay in.

A 'linspline' interpolation fits a linear spline model to the data with knots defined in knots

'nearest' finds the closest value to the new layer. You can define whether the closest value should be nearest to the top 1, or bottom 0 of the layer using int_depth

'harmonic' is similar to a linear interpolation but it uses the harmonic mean harm() using the distance in depth to each value as weights.

Multiple variables can be discretized at the same time by supplying multiple column names in param. It is also possible to use different method and controlling parameters int_depth and knots for each param. Just provide a list of settings the same length as param. If only one value is given, but multiple param the settins are reused for each parameter.

Usage

discretize_depth(
  df,
  param,
  method,
  depth_target,
  boundary_nearest = FALSE,
  boundary_average = "none",
  int_depth = 0.5,
  knots = NULL,
  ...
)

# S3 method for class 'cfp_profile'
discretize_depth(
  df,
  param,
  method,
  depth_target,
  boundary_nearest = FALSE,
  boundary_average = "none",
  int_depth = 0.5,
  knots = NULL,
  ...
)

# S3 method for class 'data.frame'
discretize_depth(
  df,
  param,
  method,
  depth_target,
  boundary_nearest = FALSE,
  boundary_average = "none",
  int_depth = 0.5,
  knots = NULL,
  id_cols = NULL,
  ...
)

Arguments

df

(dataframe) The dataframe containing the parameters to be interpolated, as well as the columns "depth", "upper" and "lower".

param

(character vector) The column names name of the parameters to be interpolated.

method

(character vector) a character (-vector) specifying the methods to be used for interpolation. Must be in the same order as param. One of

linear

boundary

linspline

nearest

harmonic

depth_target

(numeric vector or data frame) specifying the new layers. Must include n+1 depths for n target layers.

If the target layers are different for id_cols, enter a data.frame instead. This data frame must have a "depth" column, as well as well as all id_cols needed that must be at least a subset of the id_cols of the original data.

boundary_nearest

(logical) = TRUE/FALSE if it is TRUE then for target depth steps (partially) outside of the parameter boundaries, the nearest neighbor is returned, else returns NA. Default is FALSE.

boundary_average

("character) Defines what happens if the new layer contains multiple old layers. one of

none

= the deafult
the new layer is set to NA

arith

the new layer is calculated as the arithmetic mean of the old

harm

the new layer is calculated as the harmonic mean of the old

int_depth

(numeric vector) = value between 0 and 1 for 1 = interpolation takes the top of each depth step, 0.5 = middle and 0= bottom. Default = 0.5

knots

(numeric vector) = the depths at which knots for the 'linspline' method are to be placed. If this differs for the parameters, a list of numeric vectors with the same length as "param" can be provided. Cannot differ between id_cols.

...

Arguments passed on to cfp_profile

id_cols

Column names in data.frame that uniquely identify each profile.

Value

A cfp_layered_profile() data.frame with the variables upper and lower defining the layers derived from depth_target. The column depth is the middle of each layer. And all variables from param

See also

Examples

{

data("soiltemp")
library(dplyr)

dt <- soiltemp %>%
  select(site,depth) %>%
  distinct() %>%
  group_by(site) %>%
  slice_max(depth) %>%
  reframe(depth = c(depth,seq(0,-100,-10)))

discretize_depth(df = soiltemp,
                 param = "t",
                 method = "linear",
                 depth_target = dt,
                 id_cols = c(
                   "site","Date"))
}
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:ConFluxPro’:
#> 
#>     n_groups
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#> 
#> A cfp_layered_profile object 
#> id_cols: site Date 
#> 24  unique profiles 
#> 
#>       site       Date           t depth upper lower
#> 1   site_a 2021-01-01 -0.19747588 -95.0   -90  -100
#> 2   site_a 2021-01-01 -0.59242764 -85.0   -80   -90
#> 3   site_a 2021-01-01 -0.98737939 -75.0   -70   -80
#> 4   site_a 2021-01-01 -1.38233115 -65.0   -60   -70
#> 5   site_a 2021-01-01 -1.77728291 -55.0   -50   -60
#> 6   site_a 2021-01-01 -2.17223466 -45.0   -40   -50
#> 7   site_a 2021-01-01 -2.56718642 -35.0   -30   -40
#> 8   site_a 2021-01-01 -3.76136984 -25.0   -20   -30
#> 9   site_a 2021-01-01 -4.86022893 -15.0   -10   -20
#> 10  site_a 2021-01-01 -5.06453203  -5.0     0   -10
#> 11  site_a 2021-01-01 -5.97984391   2.5     5     0
#> 12  site_a 2021-02-01  0.53679849 -95.0   -90  -100
#> 13  site_a 2021-02-01  0.18280934 -85.0   -80   -90
#> 14  site_a 2021-02-01 -0.17117980 -75.0   -70   -80
#> 15  site_a 2021-02-01 -0.52516895 -65.0   -60   -70
#> 16  site_a 2021-02-01 -0.87915809 -55.0   -50   -60
#> 17  site_a 2021-02-01 -1.23314724 -45.0   -40   -50
#> 18  site_a 2021-02-01 -1.58713638 -35.0   -30   -40
#> 19  site_a 2021-02-01 -2.21903833 -25.0   -20   -30
#> 20  site_a 2021-02-01 -3.10548225 -15.0   -10   -20
#> 21  site_a 2021-02-01 -3.96855532  -5.0     0   -10
#> 22  site_a 2021-02-01 -4.35479103   2.5     5     0
#> 23  site_a 2021-03-01  2.45034545 -95.0   -90  -100
#> 24  site_a 2021-03-01  2.19007373 -85.0   -80   -90
#> 25  site_a 2021-03-01  1.92980200 -75.0   -70   -80
#> 26  site_a 2021-03-01  1.66953028 -65.0   -60   -70
#> 27  site_a 2021-03-01  1.40925855 -55.0   -50   -60
#> 28  site_a 2021-03-01  1.14898682 -45.0   -40   -50
#> 29  site_a 2021-03-01  0.88871510 -35.0   -30   -40
#> 30  site_a 2021-03-01  0.62285788 -25.0   -20   -30
#> 31  site_a 2021-03-01  0.24201891 -15.0   -10   -20
#> 32  site_a 2021-03-01 -0.24821632  -5.0     0   -10
#> 33  site_a 2021-03-01 -0.72315672   2.5     5     0
#> 34  site_a 2021-04-01  4.63736487 -95.0   -90  -100
#> 35  site_a 2021-04-01  4.69593355 -85.0   -80   -90
#> 36  site_a 2021-04-01  4.75450223 -75.0   -70   -80
#> 37  site_a 2021-04-01  4.81307092 -65.0   -60   -70
#> 38  site_a 2021-04-01  4.87163960 -55.0   -50   -60
#> 39  site_a 2021-04-01  4.93020828 -45.0   -40   -50
#> 40  site_a 2021-04-01  4.98877696 -35.0   -30   -40
#> 41  site_a 2021-04-01  4.56823988 -25.0   -20   -30
#> 42  site_a 2021-04-01  4.41830764 -15.0   -10   -20
#> 43  site_a 2021-04-01  5.01808601  -5.0     0   -10
#> 44  site_a 2021-04-01  5.31720325   2.5     5     0
#> 45  site_a 2021-05-01  8.71776729 -95.0   -90  -100
#> 46  site_a 2021-05-01  8.78971059 -85.0   -80   -90
#> 47  site_a 2021-05-01  8.86165389 -75.0   -70   -80
#> 48  site_a 2021-05-01  8.93359719 -65.0   -60   -70
#> 49  site_a 2021-05-01  9.00554049 -55.0   -50   -60
#> 50  site_a 2021-05-01  9.07748379 -45.0   -40   -50
#> 51  site_a 2021-05-01  9.14942709 -35.0   -30   -40
#> 52  site_a 2021-05-01  9.56124561 -25.0   -20   -30
#> 53  site_a 2021-05-01 10.00320297 -15.0   -10   -20
#> 54  site_a 2021-05-01 10.13542395  -5.0     0   -10
#> 55  site_a 2021-05-01 11.28144813   2.5     5     0
#> 56  site_a 2021-06-01 10.04192927 -95.0   -90  -100
#> 57  site_a 2021-06-01 10.26586735 -85.0   -80   -90
#> 58  site_a 2021-06-01 10.48980542 -75.0   -70   -80
#> 59  site_a 2021-06-01 10.71374349 -65.0   -60   -70
#> 60  site_a 2021-06-01 10.93768156 -55.0   -50   -60
#> 61  site_a 2021-06-01 11.16161964 -45.0   -40   -50
#> 62  site_a 2021-06-01 11.38555771 -35.0   -30   -40
#> 63  site_a 2021-06-01 11.73646565 -25.0   -20   -30
#> 64  site_a 2021-06-01 11.64841293 -15.0   -10   -20
#> 65  site_a 2021-06-01 10.99442969  -5.0     0   -10
#> 66  site_a 2021-06-01 14.05904033   2.5     5     0
#> 67  site_a 2021-07-01 10.36803576 -95.0   -90  -100
#> 68  site_a 2021-07-01 10.92458598 -85.0   -80   -90
#> 69  site_a 2021-07-01 11.48113620 -75.0   -70   -80
#> 70  site_a 2021-07-01 12.03768643 -65.0   -60   -70
#> 71  site_a 2021-07-01 12.59423665 -55.0   -50   -60
#> 72  site_a 2021-07-01 13.15078687 -45.0   -40   -50
#> 73  site_a 2021-07-01 13.70733709 -35.0   -30   -40
#> 74  site_a 2021-07-01 14.31162119 -25.0   -20   -30
#> 75  site_a 2021-07-01 14.00203586 -15.0   -10   -20
#> 76  site_a 2021-07-01 12.73084723  -5.0     0   -10
#> 77  site_a 2021-07-01 17.24908682   2.5     5     0
#> 78  site_a 2021-08-01  9.69338256 -95.0   -90  -100
#> 79  site_a 2021-08-01  9.90384293 -85.0   -80   -90
#> 80  site_a 2021-08-01 10.11430329 -75.0   -70   -80
#> 81  site_a 2021-08-01 10.32476366 -65.0   -60   -70
#> 82  site_a 2021-08-01 10.53522403 -55.0   -50   -60
#> 83  site_a 2021-08-01 10.74568439 -45.0   -40   -50
#> 84  site_a 2021-08-01 10.95614476 -35.0   -30   -40
#> 85  site_a 2021-08-01 12.37869429 -25.0   -20   -30
#> 86  site_a 2021-08-01 14.38298847 -15.0   -10   -20
#> 87  site_a 2021-08-01 15.75693814  -5.0     0   -10
#> 88  site_a 2021-08-01 16.95610179   2.5     5     0
#> 89  site_a 2021-09-01  7.99225807 -95.0   -90  -100
#> 90  site_a 2021-09-01  7.95784099 -85.0   -80   -90
#> 91  site_a 2021-09-01  7.92342391 -75.0   -70   -80
#> 92  site_a 2021-09-01  7.88900684 -65.0   -60   -70
#> 93  site_a 2021-09-01  7.85458976 -55.0   -50   -60
#> 94  site_a 2021-09-01  7.82017268 -45.0   -40   -50
#> 95  site_a 2021-09-01  7.78575560 -35.0   -30   -40
#> 96  site_a 2021-09-01  9.12594770 -25.0   -20   -30
#> 97  site_a 2021-09-01 10.40701863 -15.0   -10   -20
#> 98  site_a 2021-09-01 10.25435919  -5.0     0   -10
#> 99  site_a 2021-09-01 12.15748111   2.5     5     0
#> 100 site_a 2021-10-01  5.05716591 -95.0   -90  -100
#> 101 site_a 2021-10-01  5.08166485 -85.0   -80   -90
#> 102 site_a 2021-10-01  5.10616379 -75.0   -70   -80
#> 103 site_a 2021-10-01  5.13066273 -65.0   -60   -70
#> 104 site_a 2021-10-01  5.15516167 -55.0   -50   -60
#> 105 site_a 2021-10-01  5.17966061 -45.0   -40   -50
#> 106 site_a 2021-10-01  5.20415954 -35.0   -30   -40
#> 107 site_a 2021-10-01  4.95597291 -25.0   -20   -30
#> 108 site_a 2021-10-01  4.54898905 -15.0   -10   -20
#> 109 site_a 2021-10-01  4.25589354  -5.0     0   -10
#> 110 site_a 2021-10-01  4.53899437   2.5     5     0
#> 111 site_a 2021-11-01  1.69223699 -95.0   -90  -100
#> 112 site_a 2021-11-01  1.57325602 -85.0   -80   -90
#> 113 site_a 2021-11-01  1.45427505 -75.0   -70   -80
#> 114 site_a 2021-11-01  1.33529408 -65.0   -60   -70
#> 115 site_a 2021-11-01  1.21631311 -55.0   -50   -60
#> 116 site_a 2021-11-01  1.09733214 -45.0   -40   -50
#> 117 site_a 2021-11-01  0.97835117 -35.0   -30   -40
#> 118 site_a 2021-11-01  0.72205422 -25.0   -20   -30
#> 119 site_a 2021-11-01  0.27108478 -15.0   -10   -20
#> 120 site_a 2021-11-01 -0.23724116  -5.0     0   -10
#> 121 site_a 2021-11-01 -0.76927888   2.5     5     0
#> 122 site_a 2021-12-01  0.48212710 -95.0   -90  -100
#> 123 site_a 2021-12-01  0.06846827 -85.0   -80   -90
#> 124 site_a 2021-12-01 -0.34519056 -75.0   -70   -80
#> 125 site_a 2021-12-01 -0.75884939 -65.0   -60   -70
#> 126 site_a 2021-12-01 -1.17250822 -55.0   -50   -60
#> 127 site_a 2021-12-01 -1.58616705 -45.0   -40   -50
#> 128 site_a 2021-12-01 -1.99982589 -35.0   -30   -40
#> 129 site_a 2021-12-01 -2.26055623 -25.0   -20   -30
#> 130 site_a 2021-12-01 -3.00486165 -15.0   -10   -20
#> 131 site_a 2021-12-01 -4.38567065  -5.0     0   -10
#> 132 site_a 2021-12-01 -5.46154837   2.5     5     0
#> 133 site_b 2021-01-01 -0.19003417 -95.0   -90  -100
#> 134 site_b 2021-01-01 -0.57010251 -85.0   -80   -90
#> 135 site_b 2021-01-01 -0.95017086 -75.0   -70   -80
#> 136 site_b 2021-01-01 -1.33023920 -65.0   -60   -70
#> 137 site_b 2021-01-01 -1.71030754 -55.0   -50   -60
#> 138 site_b 2021-01-01 -2.09037588 -45.0   -40   -50
#> 139 site_b 2021-01-01 -2.47044422 -35.0   -30   -40
#> 140 site_b 2021-01-01 -3.09137342 -25.0   -20   -30
#> 141 site_b 2021-01-01 -4.23522906 -15.0   -10   -20
#> 142 site_b 2021-01-01 -5.66115028  -5.0     0   -10
#> 143 site_b 2021-01-01 -5.97518999   3.5     7     0
#> 144 site_b 2021-02-01  0.51180580 -95.0   -90  -100
#> 145 site_b 2021-02-01  0.17970689 -85.0   -80   -90
#> 146 site_b 2021-02-01 -0.15239202 -75.0   -70   -80
#> 147 site_b 2021-02-01 -0.48449093 -65.0   -60   -70
#> 148 site_b 2021-02-01 -0.81658984 -55.0   -50   -60
#> 149 site_b 2021-02-01 -1.14868875 -45.0   -40   -50
#> 150 site_b 2021-02-01 -1.48078766 -35.0   -30   -40
#> 151 site_b 2021-02-01 -2.30603584 -25.0   -20   -30
#> 152 site_b 2021-02-01 -3.24759795 -15.0   -10   -20
#> 153 site_b 2021-02-01 -3.81232473  -5.0     0   -10
#> 154 site_b 2021-02-01 -4.40428139   3.5     7     0
#> 155 site_b 2021-03-01  2.38760339 -95.0   -90  -100
#> 156 site_b 2021-03-01  2.17535645 -85.0   -80   -90
#> 157 site_b 2021-03-01  1.96310950 -75.0   -70   -80
#> 158 site_b 2021-03-01  1.75086256 -65.0   -60   -70
#> 159 site_b 2021-03-01  1.53861561 -55.0   -50   -60
#> 160 site_b 2021-03-01  1.32636867 -45.0   -40   -50
#> 161 site_b 2021-03-01  1.11412173 -35.0   -30   -40
#> 162 site_b 2021-03-01  0.74857063 -25.0   -20   -30
#> 163 site_b 2021-03-01  0.24864262 -15.0   -10   -20
#> 164 site_b 2021-03-01 -0.23235817  -5.0     0   -10
#> 165 site_b 2021-03-01 -0.72385380   3.5     7     0
#> 166 site_b 2021-04-01  5.07350535 -95.0   -90  -100
#> 167 site_b 2021-04-01  5.11244332 -85.0   -80   -90
#> 168 site_b 2021-04-01  5.15138129 -75.0   -70   -80
#> 169 site_b 2021-04-01  5.19031927 -65.0   -60   -70
#> 170 site_b 2021-04-01  5.22925724 -55.0   -50   -60
#> 171 site_b 2021-04-01  5.26819521 -45.0   -40   -50
#> 172 site_b 2021-04-01  5.30713318 -35.0   -30   -40
#> 173 site_b 2021-04-01  5.11761191 -25.0   -20   -30
#> 174 site_b 2021-04-01  5.00409080 -15.0   -10   -20
#> 175 site_b 2021-04-01  5.19502910  -5.0     0   -10
#> 176 site_b 2021-04-01  5.49518333   3.5     7     0
#> 177 site_b 2021-05-01  7.34627681 -95.0   -90  -100
#> 178 site_b 2021-05-01  7.76698327 -85.0   -80   -90
#> 179 site_b 2021-05-01  8.18768974 -75.0   -70   -80
#> 180 site_b 2021-05-01  8.60839621 -65.0   -60   -70
#> 181 site_b 2021-05-01  9.02910267 -55.0   -50   -60
#> 182 site_b 2021-05-01  9.44980914 -45.0   -40   -50
#> 183 site_b 2021-05-01  9.87051560 -35.0   -30   -40
#> 184 site_b 2021-05-01 10.23377393 -25.0   -20   -30
#> 185 site_b 2021-05-01 10.61665618 -15.0   -10   -20
#> 186 site_b 2021-05-01 11.07661051  -5.0     0   -10
#> 187 site_b 2021-05-01 11.21161654   3.5     7     0
#> 188 site_b 2021-06-01  9.16788014 -95.0   -90  -100
#> 189 site_b 2021-06-01  9.78426836 -85.0   -80   -90
#> 190 site_b 2021-06-01 10.40065658 -75.0   -70   -80
#> 191 site_b 2021-06-01 11.01704480 -65.0   -60   -70
#> 192 site_b 2021-06-01 11.63343301 -55.0   -50   -60
#> 193 site_b 2021-06-01 12.24982123 -45.0   -40   -50
#> 194 site_b 2021-06-01 12.86620945 -35.0   -30   -40
#> 195 site_b 2021-06-01 13.51001375 -25.0   -20   -30
#> 196 site_b 2021-06-01 14.18420826 -15.0   -10   -20
#> 197 site_b 2021-06-01 14.86137689  -5.0     0   -10
#> 198 site_b 2021-06-01 17.40943360   3.5     7     0
#> 199 site_b 2021-07-01  8.57868148 -95.0   -90  -100
#> 200 site_b 2021-07-01  9.05824260 -85.0   -80   -90
#> 201 site_b 2021-07-01  9.53780372 -75.0   -70   -80
#> 202 site_b 2021-07-01 10.01736484 -65.0   -60   -70
#> 203 site_b 2021-07-01 10.49692596 -55.0   -50   -60
#> 204 site_b 2021-07-01 10.97648708 -45.0   -40   -50
#> 205 site_b 2021-07-01 11.45604820 -35.0   -30   -40
#> 206 site_b 2021-07-01 13.82239584 -25.0   -20   -30
#> 207 site_b 2021-07-01 15.60741168 -15.0   -10   -20
#> 208 site_b 2021-07-01 14.92430917  -5.0     0   -10
#> 209 site_b 2021-07-01 17.49351653   3.5     7     0
#> 210 site_b 2021-08-01  9.34195517 -95.0   -90  -100
#> 211 site_b 2021-08-01 10.07905591 -85.0   -80   -90
#> 212 site_b 2021-08-01 10.81615666 -75.0   -70   -80
#> 213 site_b 2021-08-01 11.55325740 -65.0   -60   -70
#> 214 site_b 2021-08-01 12.29035815 -55.0   -50   -60
#> 215 site_b 2021-08-01 13.02745889 -45.0   -40   -50
#> 216 site_b 2021-08-01 13.76455964 -35.0   -30   -40
#> 217 site_b 2021-08-01 13.15905542 -25.0   -20   -30
#> 218 site_b 2021-08-01 12.37092737 -15.0   -10   -20
#> 219 site_b 2021-08-01 12.74278044  -5.0     0   -10
#> 220 site_b 2021-08-01 16.07059995   3.5     7     0
#> 221 site_b 2021-09-01  7.20725457 -95.0   -90  -100
#> 222 site_b 2021-09-01  7.39073908 -85.0   -80   -90
#> 223 site_b 2021-09-01  7.57422360 -75.0   -70   -80
#> 224 site_b 2021-09-01  7.75770812 -65.0   -60   -70
#> 225 site_b 2021-09-01  7.94119264 -55.0   -50   -60
#> 226 site_b 2021-09-01  8.12467716 -45.0   -40   -50
#> 227 site_b 2021-09-01  8.30816167 -35.0   -30   -40
#> 228 site_b 2021-09-01  9.25886753 -25.0   -20   -30
#> 229 site_b 2021-09-01 10.61045894 -15.0   -10   -20
#> 230 site_b 2021-09-01 11.59571455  -5.0     0   -10
#> 231 site_b 2021-09-01 12.50890523   3.5     7     0
#> 232 site_b 2021-10-01  6.25821774 -95.0   -90  -100
#> 233 site_b 2021-10-01  6.07276223 -85.0   -80   -90
#> 234 site_b 2021-10-01  5.88730672 -75.0   -70   -80
#> 235 site_b 2021-10-01  5.70185120 -65.0   -60   -70
#> 236 site_b 2021-10-01  5.51639569 -55.0   -50   -60
#> 237 site_b 2021-10-01  5.33094018 -45.0   -40   -50
#> 238 site_b 2021-10-01  5.14548466 -35.0   -30   -40
#> 239 site_b 2021-10-01  5.37415607 -25.0   -20   -30
#> 240 site_b 2021-10-01  5.55390660 -15.0   -10   -20
#> 241 site_b 2021-10-01  5.27060935  -5.0     0   -10
#> 242 site_b 2021-10-01  5.24469990   3.5     7     0
#> 243 site_b 2021-11-01  2.07363548 -95.0   -90  -100
#> 244 site_b 2021-11-01  1.90196454 -85.0   -80   -90
#> 245 site_b 2021-11-01  1.73029361 -75.0   -70   -80
#> 246 site_b 2021-11-01  1.55862268 -65.0   -60   -70
#> 247 site_b 2021-11-01  1.38695174 -55.0   -50   -60
#> 248 site_b 2021-11-01  1.21528081 -45.0   -40   -50
#> 249 site_b 2021-11-01  1.04360988 -35.0   -30   -40
#> 250 site_b 2021-11-01  0.70111748 -25.0   -20   -30
#> 251 site_b 2021-11-01  0.20723991 -15.0   -10   -20
#> 252 site_b 2021-11-01 -0.26720137  -5.0     0   -10
#> 253 site_b 2021-11-01 -0.65838839   3.5     7     0
#> 254 site_b 2021-12-01  0.49451890 -95.0   -90  -100
#> 255 site_b 2021-12-01  0.12542190 -85.0   -80   -90
#> 256 site_b 2021-12-01 -0.24367511 -75.0   -70   -80
#> 257 site_b 2021-12-01 -0.61277211 -65.0   -60   -70
#> 258 site_b 2021-12-01 -0.98186912 -55.0   -50   -60
#> 259 site_b 2021-12-01 -1.35096612 -45.0   -40   -50
#> 260 site_b 2021-12-01 -1.72006313 -35.0   -30   -40
#> 261 site_b 2021-12-01 -2.22915757 -25.0   -20   -30
#> 262 site_b 2021-12-01 -3.03316720 -15.0   -10   -20
#> 263 site_b 2021-12-01 -3.99209459  -5.0     0   -10
#> 264 site_b 2021-12-01 -4.51845986   3.5     7     0