Flux-gradient method

fg_flux() implements different approaches to the flux-gradient method (FGM). It takes a valid input dataset from cfp_dat() and calculates for each layer defined in cfp_layers_map().

Usage

fg_flux(x, ...)

# S3 method for class 'cfp_dat'
fg_flux(x, ...)

# S3 method for class 'cfp_fgres'
fg_flux(x, ...)

# S3 method for class 'cfp_fgmod'
fg_flux(x, ...)

Arguments

x

A cfp_dat object with all the necessary input datasets.

...

Arguments passed on to cfp_fgmod

gases

(character) A character vector defining the gases for which fluxes shall be calculated.

modes

(character) A character vector specifying mode(s) for dcdz calculation. Can be "LL","LS","EF".

LL

local linear approach: within each layer a linear model is evaluated of concentration over the depth.

LS

linear spline approach: A linear spline is fitted over the complete profile with nodes at the layer intersections.

EF

exponential fit approach: An exponential function of form y=a+b*x^c is fit of concentration against depth. Using the first derivative of that function the concentration gradient is evaluated for each layer.

DA

exponential fit approach: An exponential function of form y=a+b*(1-exp(-a*x)) is fit of concentration against depth. Using the first derivative of that function the concentration gradient is evaluated for each layer. From Davidson (2006).

param

(character) A vector containing the the parameters of soilphys, for which means should be calculated, must contain "c_air" and "DS", more parameters may help interpretation.

funs

(character) A vector defining the type of mean to be used for each parameter in param. One of "arith" or "harm".

References

DAVIDSON, E. A., SAVAGE, K. E., TRUMBORE, S. E., & BORKEN, W. (2006). Vertical partitioning of CO2 production within a temperate forest soil. In Global Change Biology (Vol. 12, Issue 6, pp. 944–956). Wiley. https://doi.org/10.1111/j.1365-2486.2005.01142.x

See also