Plot profiles

Plot vertical soil profiles of ConFluxPro objects using ggplot. This is mainly intended for diagnostic purposes and better understand the underlying data.

Supported objects:

cfp_pfres

Displays TPS, SWC and AFPS, as well as production and measured and modelled gas concentrations.

cfp_fgres

Displays TPS, SWC and AFPS, as well as the measured concentration profile, and concentration gradients for each layer.

cfp_soilphys

Displays TPS, SWC and AFPS, as well as values of Ds and Temperature.

cfp_gasdata

Displays the concentration profile.

cfp_layers_map

Displays the layer names, pmap and layer_couple, as well as the allowed production range.

Usage

plot_profile(x)

Arguments

x

A cfp_pfres, cfp_fgres model result, or a cfp_soilphys, cfp_gasdata or cfp_layers_map object

Value

A ggplot2 plot with facets for each distinct profile. If more than 20 profiles are plotted a message is sent because this can take a long time.