prod_optim

This is the optimizer-function that is minimized for the inverse, production based model. It takes as input a vector of the influx, as well as the values of the production to be optimized.

This function is embedded in pro_flux() and is not intended to be used manually.

Usage

prod_optim(
  X,
  height,
  DS,
  D0 = NA,
  C0,
  pmap,
  cmap,
  conc,
  dstor = 0,
  zero_flux = TRUE,
  F0 = 0,
  known_flux = NA,
  known_flux_factor = 0,
  DSD0_optim = FALSE,
  layer_couple,
  wmap,
  evenness_factor
)

Arguments

X

(numeric vector) specifying the production rates to be optimized

height

(numeric vector) giving the height of each step

DS

(numeric vector) giving the DS of each step

D0

RESERVED FOR FUTURE EXPANSION

C0

(numeric) The concentration at the bottom of the lowermost step.

pmap

(integer vector) assigning a production from X to each step

cmap

(integer vector) assigning the modeled concentrations to the observed concentrations as there can be multiple observations per depth

conc

(numeric) the observed concentrations (in the same unit as the modelled concentrations).

dstor

RESERVED FOR FUTURE EXPANSION

zero_flux

(logical) Applies the zero-flux boundary condition(TRUE)? If FALSE, the first value in X. represents the incoming flux to the lowest layer.

F0

(numeric) flux into lowest layer.

known_flux

RESERVED FOR FUTURE EXPANSION

known_flux_factor

RESERVED FOR FUTURE EXPANSION

DSD0_optim

RESERVED FOR FUTURE EXPANSION

layer_couple

(numeric vector) A vector defining the weights that bind the different layers together. If all is zero, no penalisation for stark differences between the optimized production rates of adjacent layers takes place

wmap

(numeric) A vector defining the weights of the different concentration measurements in the RMSE calculation.

evenness_factor

(numeric) Defines strong should stark differences between the production rates and very small production rates be penalized.

Value

A modified RMSE root mean square error of the modeled and measured concentration.

See also