Extract efflux rates

Calculate or extract the soil/atmosphere efflux from cfp_pfres or cfp_fgres model results.

Usage

efflux(x, ...)

# S3 method for class 'cfp_pfres'
efflux(x, ...)

# S3 method for class 'cfp_fgres'
efflux(x, ..., method = "lm", layers = NULL)

# S3 method for class 'cfp_altres'
efflux(x, ...)

Arguments

x

A cfp_pfres or cfp_fgres model result, or a cfp_altres.

...

Arguments passed to methods.

method

Method(s) used to interpolate the efflux at the top of the soil from partial fluxes within the soil. One of

top

Use the flux in the topmost model layer.

lm

A linear model where each partial flux is centered in the respective layer and the model is evaluated at the top of the soil.

lex

Linearly exterpolate using fluxes of two layers in the soil.

layers

Vector of two integers selecting the layers for the lex method. Layers are indexed from 1 (topmost) to the number of layers used in the flux calculation.