Create a run plan for sobol indice calculation

Modify an existing cfp_run_map for sobol indice estimation or create a new one from scratch.

Usage

sobol_run_map(x, ...)

# S3 method for class 'cfp_dat'
sobol_run_map(x, ...)

# S3 method for class 'cfp_run_map'
sobol_run_map(x, ...)

Arguments

x

Either an object of class cfp_run_map created by a call to cfp_run_map() with method = 'random', or a cfp_pfres or cfp_fgres model result.

...

Arguments passed on to run_map

params

A named list of numeric vectors. Names indicate column names in soilphys, vectors either distinct values (method permutation) or limits (method random).

type

A vector of length param indicating what the values in params represent. One of

abs

Absolute values that are applied as-is.

factor

Factors to be multiplied with the original values.

addition

Factors to be added to the original values.

method

Either 'random', where a random value is chosen within the bounds set in params or 'permutation', where every permutation of the values in params is added.

n_runs

Integer value of the number of alterations to be done for method = 'random'.

layers_different

Should layers from layers_map be changed individually? If TRUE this allows for different changes at different depths.

layers_from

(character) If layers_different is TRUE, from which source should the layers be created? One of:

layers_map

(default) Use the layers that are defined in layers_map.

soilphys

Use the layers as defined in soilphys

layers_altmap

Use the layers as defined in the provided layers_altmap object.

layers_altmap

An optional layers_map created using layers_map() that defines the layers to be used if layers_different = TRUE.

topheight_adjust

(logical) If the proposed change in topheight is larger than the highest layer in soilphys, should the limits be automatically adjusted per id_cols indivdually? Default is FALSE, which leads to an error in that case.

See also